Move fuzz tests to their corresponding type test file.
